+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+### Added
+
+- Added support for repeat timers via `:TimerRepeat`.
+
 ## [v0.2.0](https://github.com/epwalsh/pomo.nvim/releases/tag/v0.2.0) - 2023-11-30
 
 ### Added